server/requirements.txt
anyio==3.7.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
ecdsa==0.18.0
fastapi==0.96.1
fastapi-mqtt==1.0.9
gmqtt==0.6.12
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
oauthlib==3.2.2
packaging==23.1
pluggy==1.0.0
pyasn1==0.5.0
pydantic==1.10.9
pytest==7.3.2
python-dotenv==1.0.0
python-jose==3.3.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a34
sqlmodel==0.0.8
starlette==0.27.0
typing_extensions==4.6.3
urllib3==2.0.3
uvicorn==0.22.0
